Switch to common string sizes
HW objects and their Python-side definitions use statically-sized
string for various fields in their config and status structures.

To avoid any sizing issues between the C and Python sides, use
common definitions for the string sizes. These are defined by the
core extension module and are added to it as integer constants,
which are then used by the Python code to define the size of the
ctypes structure fields.