Do a better job at parsing object state when looking for pin names
The Klipper frontend builds a pin-to-object map from all of the
emulator objects in order to be able to match pin names to objects
when Klipper starts configuring the emulator.

When doing so, the frontend did not consider all possible settings
where the string "pin" appears, resulting is "invalid" entries in
the pin map.