Enable virtual object command submission and state querying from HW objects
Up to now, command submission and querying state has been only between
HW objects and from the frontend to the emulator objects.

This commit enables HW object to query and submit commands for virtual
objects, as well.

To do this, a C header file needs to be generated that will allow HW
objects to work with virtual object commands and state. Just like for
commands and state of HW objects, the data is converted to/from C
structure from/to Python ctypes structure, except this is now done in
the other direction, respectively. This requires that virtual objects
also define a set of ctypes structures that describe their commands'
arguments and state.