Switch Klipper move queue to a "global" queue
Klipper firmware uses a single move queue for the entire system
and tracks "moves" at a global scope.

The emulator, on the other hand used a separate queue for each
stepper motor. In practice, this didn't matter since the Klipper
host ensures that it does not submit more moves than the size of
the queue. However, in order to accumately emulate the Klipper
firmware, a single queue for all objects should be used.