Rename Vortex executable to vortex_run.py
Commit a80f309b2670 ("Rename `vortex_emulator.py` to `vortex.py`)
renamed the Vortex executable to 'vortex.py', which at the time
seemed much more appropriate. This worked with the development
build but does not work when installed in a virtual environment.

When running 'vortex.py' from the installed wheel in the virtual
environment, the following import exception is produced:

  Traceback (most recent call last):
    File "/home/dharalanov/tmp/vortex-env/bin/vortex.py", line 22, in <module>
      import vortex.emulator
    File "/home/dharalanov/tmp/vortex-env/bin/vortex.py", line 22, in <module>
      import vortex.emulator
  ModuleNotFoundError: No module named 'vortex.emulator'; 'vortex' is not a package

because both the executable and the Vortex module are named the
same. To avoid this, the executable script is renamed to
'vortex_run.py'. Not as nice as 'vortex.py' but, unfortunately,
necessary.