Initialize object event caches
When creating object caches, the object cache creation API allows
for multiple calls to object_cache_create() with the same cache
pointer. If the pointer is NULL, the cache is created. Otherwise,
the cache's reference count is incremented.

Some objects caches are global and used by all objects created. If
the cache pointer is not initialized and happens to be set to some
garbage value, object_cache_create() might attempt to use the value
as a valid cache and dereference it when incrementing the reference
count.