Re-implement logging for better performance
Logging in the emulator was done using the Python
`logging` module. This module provides a very
flexible and configurable logging facility.

However, one issue that existed with the emulator
using it is that all logging should be done using
the `logging` module. This means that even HW
objects had to use it. In order for the HW objects
to use the Python module, all logging calls had to
be routed through the CPython API, which meant
taking the Python GIL. Some simple measurements
were showing that, on average, logging alls from HW
objects were taking in the order of 1ms per call.
This is very slow when the emulator is trying to
keep as precise timing as possible.

This commit re-implements logging within the Vortex
emulator by moving it to a C library which can be
used by the HW objects directly. For Python code,
the library is wrapped by a Python module through
CFFI.

The new implementation is not as feature-rich as the
Python one but it does emulate the most important
feature - hierarchical naming - which allows for
message filtering.