Determine if toolhead is at origin based on all axes
It is incorrect to only look at the X, Y, and Z axis in order
to detemine whether the toolhead is at it's origin. It works
for 3 axis machines but not for ones with more.

Instead of looking only at the 3 primary axes, look at all the
axes to which the toolhead is connected.