Rework probe trigger logic
The way the probe trigger logic works was using a fuzz distance
that was randomly generated within a pre-define range (through
the configuration file). If the probe axis position was below
this fuzz value, the probe will be triggered.

The intention was to simulate something close to a real-world
triggering mechanism where the probe will not trigger exactly
at the same axis position.

This was proving unreliable (especially for the Klipper frontend)
due to the fact that the fuzz value could be in the thousands of
a millimeter. Often the axis could not achieve such precision in
its movement, leading to the probe not triggering.

This commit reworks how the probe triggering works. Instead of
using a randomly generated, high precision value, the trigger
point is replaced by a single limit value. If the axis position
is below that value, the probe triggers.

Techincally, the random fuzz value logic could still be kept by
providing a lower and upper limit of the fuzzed value that is
more realistic. However, for now this is simpler.