Differentiate between control and position axis for toolheads
The toolhead object needs to differentiate between the axes that
influence the toolhead's position and the axes in which the toolhead's
position is reported.

A perfect example of this is Delta machines. On such machines, the
toolhead's position is controlled by the A, B, and C axes. However,
the toolhead moves in the X, Y, and Z axes.

So now, the toolhead object has two separate settings - 'axes' and
'attachment'. 'axes' is the set of axes in which the toolhead's
move and 'attachment' is the set of axes which control the toolhead's
position.