Only cleanup the controller if it's instantiated
The Controller class can throw and exception during the `__init__`
call if the core fails to create and initialize all of the HW
objects.

In that case, the Emulator's __del__ method will be called when
the emulator is destroyed on program exit. That call is supposed
to clean up the controller but if the Controller is not
instantiated, that results in a secondary exception.