Rework/improve test framework
This is the first in a series of commits improving/fixing/revamping
the test framework.

This commit improves that structure of the framework. Previously,
the framework served as a common "helper" to all of the tests.
Now, the frameworks is responsible for running the tests. A lot of
the code that was in the main test executable and in the individual
tests has been moved into the framework. With this, the framework
will find all tests applicable to a frontend, resolve test
dependencies, execute the tests, and display results.

As for the individual tests, a new `testutils` module is provided
that contains common code, which hooks into the framework. The
main thing in the `testutils` module is a function decorator that
each test function is decorated with. It dynamically replaces some
of the code in the tests and sets some function properties which
are then used to identify test functions.

Finally, tests are organized per frontend, where all tests for a
particular frontend are in their own directory.