Limit the precision when comparing temperature to target
float type can support up to 7 decimal places of precision. Using
such precisions when comparing current temperature to the target
temperature is not necessary and could lead to an infinite loop
since it's possible for the temperature to never converge on a
single value.

Therefore, only use three decimal places when comparing temperatures.
That precisions should be enough because we don't expect any client
or frontend to care about more than that.