Correct the stepper step calculation
Stepper objects calculated the current step count based on the
acculated steps for the current move minus the current step
count. This was done so the current step count was incremented
only by the delta steps between update cycles.

However, using this method lead to invalid calculations when
another move was submitted. The acculated steps for the new move
would be reset to 0 when the previous move finished. This
resulted to the current step count instantly being reset to 0
since the calculation subtracted that value from 0.

Instead, calculate the delta steps based on the previous value
of the acculated steps for the move and increment the current
steps by that value removing the need to subrate the current
step count.