Improve layout of monitor application UI
This commit improves the layout of the monitor.py application
UI in several ways:

1. Allow user to select which object klasses to show
With increasing the number of objects supported by the emulator,
the monitor.py application was getting a little hard to use.

The window contained a lot of information, which could be somewhat
overwhelming.

To improve this, a set of checkbuttons are now available which can
be used to remove information about each of the object klasses as
needed. This way only the objects the user is interested in will
be displayed.

2. Place control buttons along all emulation state widgets
The control buttons were below the runtime and clock state display
widgets, which was not how they should have been layed out. Placing
them on the side of ALL emulation state widgets give a better
impression that they control the entire emulation, not just the
object state.

3. Re-organize object state and commands into grids
Previously, object state and command widgets were layed out using
Gtk boxes. While this worked, it made the layout more confusiong
as properties were not vertically lined up. By switching the
layout to using Gtk grids, the individual properties are now
visually lined up and easier to read/understand.