Add a process lock file
There was nothing stopping multiple instances of the emulator from
running at the same time. This would not work because all instances
would be using the same control pipe and would interfere with each
other.

This commit adds a file that each emulator instance attempts to open
and lock. If the lock succeeds, the emulator instance starts. If not,
an error message is diplayed and the instance exits.