Fix a race condition with controller command completion
The controller keeps a mapping between submitted command IDs and
the object for which the command was submitted in order to be
able to look up the object on command completion. This is needed
so any returned data can be mapped to the correct data structure.

There was a race condition where the command could complete prior
to the entry in the mapping having been entered, which results in
an invalid object being returned from the lookup.