Don't wait for command completion in SPI helper's send() call
The SPI helper's send() sends data or commands to objects that
use the SPI protocol. The protocol, itself, does not provide for
a acknowledgement of successful transfer so there is no point
in waiting for the command to complete. This also improves
performance of the Klipper frontend if a DISPLAY (currently, the
only SPI client object) is defined.