Use a kernel module for time control
Time control from userspace is not very precise. The combination of
`clock_gettime()` and `nanosleep()` calls adds a lot of latency,
which results in the controller clock granularity being very coarse.

In some tests, a call to `nanosleep(1)` averages at ~75k ns per
call. With such latency, a controller running in the MHz range will
seek a jump in the its clock by 75 ticks.

The kernel module serve to provide timing by using a combination of
`read()` and `ioctl()` system calls. The emulator uses `ioctl()` to
configure and control the timing and `read()` to get timing data
(runtime and clock ticks). To prevent the emulator from consuming
100% of a CPU core, the kernel timing can be configured to sleep for
some time before returning from the `read()` call.

The `read()` system (which provides the data) has latency in the
order of 700 ns when there is no sleep configured. Furthermore, with
a sleep time of 1000ns, the average runtime jump between calls is
~5000ns. This results in a tick jump of ~64 ticks for a controller
running at 120MHz as opposed to 903 for the same controller using
`nanosleep()` and `clock_gettime()`.