Allow multiple logging streams
This commit changes how logging output is handled. Instead of
having one output stream (stderr or a file), it is now possible
to add multiple stream with different logging levels.

The normal logging streams don't have any special handling in the
low level logging code any more. They are just any normal stream.

This allows for better usability/debuggability - the console can
display messages with a lower debug level so it is not overwhelmed
with output while the full debug output could be going to a file.