Unpause time control thread if paused during clean-up
If the time control thread is paused during clean-up, other
threads might be blocked on the `futex()` calls, in which
case they would not be able to exit. This causes the emulation
to hang.

To deal with this issue, explicitly resume the time control
thread if paused when `core_threads_stop()` is called.

In addition, store a pointer to the time_control thread in the
global thread data structure for easy reference. The reference
is needed in multiple functions which used to traverse through
all threads to find it.