Add support for ctypes.POINTER fields in object structures
This commit adds support for ctypes.POINTER field types in the
object structures to the C header generation tool.

It is important to note that there is no good way to determine
the size of the value(s) which the pointer points to unless the
Python set of values have a pre-defined size, in which case it
would be better to use ctypes Arrays. Otherwise, the ctypes
helper functions which fill out and parse the structure will
end up going into an infinite loop reading values and eventually
crash the emulator.