Rename ObjectTypes to ObjectKlass
ObjectTypes was an enumeration that lists all of the "types" of
object (both HW and virtual). However, through out the code the
values were assigned to variables called "klass" (or forms of it)
since "type" is a built-in function in Python.

Renaming ObjectTypes to ObjectKlass makes things consistent and
removes the issues with using "type" - all variables can now use
"klass" in their name.

The change is pretty large since the enumeration is very widely
used in all parts of the code base.