Complete stepper command from command handlers
With the exception of the "use_pins" command, tepper objects
would save the submitted command after the command handler
had run and then complete it from the update function.

For the most part, this work. However, for the "use_pins" command
it would create a situation where the command would be saved after
the handler had already attempted to complete it.

Then, when the stepper object is reset, the same command ID would
be used to complete any outstanding commands.

This commit fixes this issue by reworking how command completion
works for stepper objects. Commands are completed directy by any
handler that does not require the update callback to be called in
order to make progress.