Fix resetting of virtual objects
Resetting of virtual objects was broken if the controller's
reset() method was called with an empty list of objects to be
reset. The intent was to reset all objects in that situation.

The reset() method first attempts to reset all virtual objects,
then it calls into the core to reset all HW objects.

The core deals with an empty list as intended. It resets all
HW objects. However, for the virtual objects nothing was being
reset since the list of objects passed into the method was
empty.