Allow tracking of arbitrary objects/properties in emulator_track.py
Previously, emulator_track.py only tracked the toolhead position and
heater temperatures.

However, it would be more useful to allow the tool to track any
property of any object and generate a graph of the values.

For now, the generated graphs are only scatter plots with the
exception of toolhead position.