Split controllers into boards and chips
Controllers are techinically two separate parts - the SoC and the
board. The two parts define separate properties of the controller.
For example, the architecture and clock frequency is a property
of the SoC and the pin names are a property of the board.

This commit splits the controllers into two separate parts - the
board class and the chip class. A board class, which makes up the
controller inherits from a chip class and the base Controller
class. This way, there can be boards which implement different
SoC (for example, different revision of the same board with
different SoC).