Convert object_defs.py to auto-generated file
In order for the Python side of the emulator code to be able to
communicate with the HW objects, structures used the HW objects
for configuration, state, command arguments, etc. have to be
replicated in Python using the `ctypes` module.

However, it is possible for the Python structure definitions to
be auto-generated by parsing the C HW object code.

This is what this change does. It extendes the `tools/auto_gen.py`
tool to parse all of the HW object files and create a new
`object_defs.py` file containing the necessary structures.

This new method does require changes in development processes. Now,
HW object structures have to have specific name patterns, which have
been outlined in the documentation.