Improve emulator event handling
There could be multiple subscription to the same event of the same
object. These subscription need to all be triggered.

The core's event handling on the Python side had a bug where only a
single handler for a particular object/event combination could be
registered.

This commit fixes this by reworking how event subscriptions are
handled. Now, each subscription will return a "token" - an opaque
unique value to identify the subscription. Using tokens allows to
uniquely identify each subscription, regarless for which event and
object they are.