Better toolhead posistion detection and even triggering
The algorithm for detecting of the toolhead is at its origin point
was flawed in that assume that the axis positions will reach precisely
0.0. However, this may not always be the case. The ending axis
position could be a fraction of a millimeter away from 0.0, in which
case the toolhead will never reach the origin point.

Instead of assuming that the axes will reach 0.0, compute the axes
positions up to a certain precision and use that value to determine
the toolhead position.

In addition, ensure that the TOOLHEAD_ORIGIN event is triggered only
once per reaching the origin point.