Detangle command processing
The way commands were submitted and processed was to
convoluted.

Commands for HW objects would go through the following
path:

    frontend -> queue -> controller -> core -> object

Completions would come back like this:

    object -> core -> controller -> frontend

This is the intended path. Once the command is queue by
the frontend, it will be picked up and sent to the
controller, which would encode the options and send it
to the core. The core would then dispatch to the HW object.

Completions would follow a similar path (except the queue).
A command completion would be sent by the core to the
controller, any data would be decoded, and then the completion
is sent to the frontend.

Commands for virtual objectes followed seemingly similar
path:

    frontend -> queue -> controller -> virtual object

The completions then would:

    virtual object -> controller -> frontend

However, there were a couple of big issues with how this
was done:

    1. Virtual object command completions don't need to
       go through the controller.
    2. Virtual object command handling required a lot of
       code to separate their handling from the handling
       of HW object commands. For example, handling
       command completions in the controller required a
       separate completions method.

This commit fixes this by making the virtual object commands
follow the same path as HW objects when submitted. Since
virtual objects have to be registered with the core, it is
possible for the core to send the commands to them.

On the completion path, virtual object completions now use
the same controller methods as HW objects allowing for the
removal of all the "duplicated" code.