Move command completion to command queue
The base frontend had a somewhat complicated mechanism for
tracking outstanding commands in order to process completions
and allow for the callers to wait for command completions.

This required managing multiple lists with their associated
locks. In order to avoid any completions triggering prior to
the frontend being able to insert the command into the
appropriate list.

However, it is much simpler to move all of that logic to the
command queue. The queue already knows about all submitted
commands since frontends submit all commands through it. It
can also managed completions using a single since there is
no possible way for the command to complete prior to it
being de-queued. This makes the entire management and frontend
simpler.